requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-admin==1.2.4
django-filter==1.0.4
django-nose==1.4.5
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Logentries==0.17
MarkupSafe==1.0
nose==1.3.7
openapi-codec==1.3.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
coverage==4.4.1
celery==4.1.0
faker==0.8.4
pendulum==1.3.0
attrs==17.2.0
gevent==1.2.2
psycogreen==1.0
gunicorn==19.7.1
ipython==6.2.1